+++ src/lib/routes.js
@@ -6,8 +6,8 @@ import getHash from './utils/getHash.js';
 import wall from './page/wall.js';
 import { removeHashes, sendRoute } from './utils/clearHash.js';
 import eventButtonContinue from './utils/eventButtonContinue.js';
-import { displayUserData } from '../firebase/displayUserData.js';
-import { eventLoginButton } from './utils/loginButton.js';
+import { displayUserData } from '../firebase/authenticationFirebase.js';
+import { eventLoginButton } from '../firebase/authenticationFirebase.js';
 
 const containerPage = document.getElementById('contentPageId');
