+++ src/lib/utils/signOut.js
